Add tokenizer.cc, update http_request, CMakeLists.txt
tokenizer is supposed to read from a buffer and write out characters to
form lines. The idea is since HTTP streams don't all show up at once,
use the tokenizer to get lines from the stream, and keep track of
whether more data needs to be loaded to complete a line. This process is
still pretty manual, hopefully will figure out a nicer way to do it in
the future. Check http_request.cc for current usage.

http_request now takes a SSL_CTX so it can create TLS connections when
needed.