Update progess, add stream and http_response
Work in progress, does not compile completely at the moment.

stream is a little abstraction over a raw cmd::socket. You can ask for
lines (delimited by CRLF or LF), or for a specific amount of bytes. It
handles buffering of more data when it needs more to meet the requested
amount of data.

http_response will represent an http response. It takes a cmd::stream to
parse. It will look at headers to see how to parse the body. It should
only read the data needed to process the response -- i.e. it wont block
trying to read when the connection is kept alive. It will handle chunked
messages, content-length messages, and connection closed messages.