Use regex for URL parsing
http_response will no longer read any body if the status code is 204.

URL parsing now does some sane checks through the use of a regular
expression. It probably doesn't accept all valid URLs, but it accepts
many. Added option for port to be specified with :<portno> as a suffix
to the host.

Tested HTTP pipelining. http_request and http_response seemed to behave
correctly, but this was a very limited test.
Every http_request::connect() should be followed by a
http_request::response() to get the response in the order called by
the connects.