Update project, fix code,
Update CMakeLists.txt for http_pool. Move headers under HEADER_FILES
variable. Now the code unrelated to main is built into a shared library
that main links to.

http_response now notifies http_pool when connections are closed
(Connection header).

http_response will retry 2 times when an exception is thrown when
sending a request or attempting to connect to the host. The reason for
this is because the connection might have timed out due to a long period
of time between requests, or something terminated the connection.

plain_socket now uses the up-to-date method of finding host by name
(getaddrinfo) instead of the deprecated method used before
(gethostbyname). So far it is much faster at lookups.