Update socket interface, fix code
socket interface now exposes the socket's file descriptor.

tls_socket no longer requires an SSL_CTX* to be passed in for its
constructor. Instead, it grabs it from cmd::ssl_manager::get_context()

Remove exception throwing from close methods of various sockets. This is
to protect destructors so they finish gracefully.

In ssl_manager there were instances that the SSL_CTX could be leaked in
the event that an exception is thrown during construction. Now the
context is freed before throwing the exception.