Refactor code, optimize imports, change interfaces
socket's now return ssize_t like the underlying calls.

http_request no longer parsers a URL to figure out the host, protocol,
port, and resource. Now it just takes a cmd::stream which it assumes is
already connected. In the future I hope to add a function to do this
parsing but have it support other protocols like WebSocket (ws, wss).
The old code for parsing is now in resource_parser, but it is unused.

websocket now makes use of http_pool to get a connection and uses
http_request to send the upgrade request.

Move bind_server_socket to new file (server_socket.cc), previously in
tcp_server.