Update udp_socket.{cc,h}
Mostly the same behavior as before, but now there is an inet_addr class
to represent where a packet is going to, or where a packet is coming
from. cmd::udp_socket can send packets, but it cannot recv them because
it is unbound. To receive datagrams, use the bound_udp_socket.

There is also 2 handy methods in inet_addr to determine the textual
representation of the IP address and to get the underlying port number.