Add default constructor for cmd::stream
cmd::stream now uses a vector<char> instead of an array to better
support moves when streams are copied. I'm still thinking about
removing the copy assignment and constructors for streams because it
might lead to copied streams to think the underlying stream source is in
a different state that it appears.

For example if you create a stream from a tcp socket and make a few
http requests, then make a copy of the stream, the copied stream will
have the same buffer as the first. When it goes to read more than is
buffered, it will read from the source and possibly get nonsense
information.