Change websocket connect method
Instead of establishing the TCP connection in the constructor, delay it
until the connect method is called. Doing this means you can pass the
url directly to the connect method without having to store the resource
or url to connect to.

There are 2 methods for upgrading to a WebSocket: direct websocket
encoded url (e.g. ws://www.example.com/resource) or by specifying the
host, port, resource, and if it is a secure connection.